+++ packages/core/src/PasteRule.ts
@@ -21,8 +21,12 @@ export type PasteRuleMatch = {
   data?: Record<string, any>
 }
 
-export type PasteRuleFinder = RegExp | ((text: string, event?: ClipboardEvent) => PasteRuleMatch[] | null | undefined)
+export type PasteRuleFinder = RegExp | ((text: string, event?: ClipboardEvent | null) => PasteRuleMatch[] | null | undefined)
 
+/**
+ * Paste rules are used to react to pasted content.
+ * @see https://tiptap.dev/guide/custom-extensions/#paste-rules
+ */
 export class PasteRule {
   find: PasteRuleFinder
 
@@ -33,8 +37,8 @@ export class PasteRule {
     commands: SingleCommands
     chain: () => ChainedCommands
     can: () => CanCommands
-    pasteEvent: ClipboardEvent
-    dropEvent: DragEvent
+    pasteEvent: ClipboardEvent | null
+    dropEvent: DragEvent | null
   }) => void | null
 
   constructor(config: {
@@ -43,9 +47,9 @@ export class PasteRule {
       can: () => CanCommands
       chain: () => ChainedCommands
       commands: SingleCommands
-      dropEvent: DragEvent
+      dropEvent: DragEvent | null
       match: ExtendedRegExpMatchArray
-      pasteEvent: ClipboardEvent
+      pasteEvent: ClipboardEvent | null
       range: Range
       state: EditorState
     }) => void | null
@@ -58,7 +62,7 @@ export class PasteRule {
 const pasteRuleMatcherHandler = (
   text: string,
   find: PasteRuleFinder,
-  event?: ClipboardEvent,
+  event?: ClipboardEvent | null,
 ): ExtendedRegExpMatchArray[] => {
   if (isRegExp(find)) {
     return [...text.matchAll(find)]
@@ -97,8 +101,8 @@ function run(config: {
   from: number
   to: number
   rule: PasteRule
-  pasteEvent: ClipboardEvent
-  dropEvent: DragEvent
+  pasteEvent: ClipboardEvent | null
+  dropEvent: DragEvent | null
 }): boolean {
   const {
     editor, state, from, to, rule, pasteEvent, dropEvent,
@@ -154,6 +158,16 @@ function run(config: {
   return success
 }
 
+const createClipboardPasteEvent = (text: string) => {
+  const event = new ClipboardEvent('paste', {
+    clipboardData: new DataTransfer(),
+  })
+
+  event.clipboardData?.setData('text/html', text)
+
+  return event
+}
+
 /**
  * Create an paste rules plugin. When enabled, it will cause pasted
  * text that matches any of the given rules to trigger the rule’s
@@ -164,8 +178,47 @@ export function pasteRulesPlugin(props: { editor: Editor; rules: PasteRule[] }):
   let dragSourceElement: Element | null = null
   let isPastedFromProseMirror = false
   let isDroppedFromProseMirror = false
-  let pasteEvent = new ClipboardEvent('paste')
-  let dropEvent = new DragEvent('drop')
+  let pasteEvent = typeof ClipboardEvent !== 'undefined' ? new ClipboardEvent('paste') : null
+  let dropEvent = typeof DragEvent !== 'undefined' ? new DragEvent('drop') : null
+
+  const processEvent = ({
+    state,
+    from,
+    to,
+    rule,
+    pasteEvt,
+  }: {
+    state: EditorState
+    from: number
+    to: { b: number }
+    rule: PasteRule
+    pasteEvt: ClipboardEvent | null
+  }) => {
+    const tr = state.tr
+    const chainableState = createChainableState({
+      state,
+      transaction: tr,
+    })
+
+    const handler = run({
+      editor,
+      state: chainableState,
+      from: Math.max(from - 1, 0),
+      to: to.b - 1,
+      rule,
+      pasteEvent: pasteEvt,
+      dropEvent,
+    })
+
+    if (!handler || !tr.steps.length) {
+      return
+    }
+
+    dropEvent = typeof DragEvent !== 'undefined' ? new DragEvent('drop') : null
+    pasteEvent = typeof ClipboardEvent !== 'undefined' ? new ClipboardEvent('paste') : null
+
+    return tr
+  }
 
   const plugins = rules.map(rule => {
     return new Plugin({
@@ -212,45 +265,45 @@ export function pasteRulesPlugin(props: { editor: Editor; rules: PasteRule[] }):
         const isPaste = transaction.getMeta('uiEvent') === 'paste' && !isPastedFromProseMirror
         const isDrop = transaction.getMeta('uiEvent') === 'drop' && !isDroppedFromProseMirror
 
-        if (!isPaste && !isDrop) {
+        // if PasteRule is triggered by insertContent()
+        const simulatedPasteMeta = transaction.getMeta('applyPasteRules')
+        const isSimulatedPaste = !!simulatedPasteMeta
+
+        if (!isPaste && !isDrop && !isSimulatedPaste) {
           return
         }
 
-        // stop if there is no changed range
+        // Handle simulated paste
+        if (isSimulatedPaste) {
+          const { from, text } = simulatedPasteMeta
+          const to = from + text.length
+          const pasteEvt = createClipboardPasteEvent(text)
+
+          return processEvent({
+            rule,
+            state,
+            from,
+            to: { b: to },
+            pasteEvt,
+          })
+        }
+
+        // handle actual paste/drop
         const from = oldState.doc.content.findDiffStart(state.doc.content)
         const to = oldState.doc.content.findDiffEnd(state.doc.content)
 
+        // stop if there is no changed range
         if (!isNumber(from) || !to || from === to.b) {
           return
         }
 
-        // build a chainable state
-        // so we can use a single transaction for all paste rules
-        const tr = state.tr
-        const chainableState = createChainableState({
-          state,
-          transaction: tr,
-        })
-
-        const handler = run({
-          editor,
-          state: chainableState,
-          from: Math.max(from - 1, 0),
-          to: to.b - 1,
+        return processEvent({
           rule,
-          pasteEvent,
-          dropEvent,
+          state,
+          from,
+          to,
+          pasteEvt: pasteEvent,
         })
-
-        // stop if there are no changes
-        if (!handler || !tr.steps.length) {
-          return
-        }
-
-        dropEvent = new DragEvent('drop')
-        pasteEvent = new ClipboardEvent('paste')
-
-        return tr
       },
     })
   })