+++ packages/core/src/helpers/getMarksBetween.ts
@@ -26,7 +26,7 @@ export function getMarksBetween(from: number, to: number, doc: ProseMirrorNode):
       })
   } else {
     doc.nodesBetween(from, to, (node, pos) => {
-      if (!node || node.nodeSize === undefined) {
+      if (!node || node?.nodeSize === undefined) {
         return
       }