+++ packages/extension-code-block-lowlight/src/code-block-lowlight.ts
@@ -3,10 +3,23 @@ import CodeBlock, { CodeBlockOptions } from '@tiptap/extension-code-block'
 import { LowlightPlugin } from './lowlight-plugin.js'
 
 export interface CodeBlockLowlightOptions extends CodeBlockOptions {
+  /**
+   * The lowlight instance.
+   */
   lowlight: any,
+
+  /**
+   * The default language.
+   * @default null
+   * @example 'javascript'
+   */
   defaultLanguage: string | null | undefined,
 }
 
+/**
+ * This extension allows you to highlight code blocks with lowlight.
+ * @see https://tiptap.dev/api/nodes/code-block-lowlight
+ */
 export const CodeBlockLowlight = CodeBlock.extend<CodeBlockLowlightOptions>({
   addOptions() {
     return {