+++ packages/extension-text/src/text.ts
@@ -1,5 +1,9 @@
 import { Node } from '@tiptap/core'
 
+/**
+ * This extension allows you to create text nodes.
+ * @see https://www.tiptap.dev/api/nodes/text
+ */
 export const Text = Node.create({
   name: 'text',
   group: 'inline',